requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
DateTime==4.3
Flask==2.0.1
future==0.18.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
line-bot-sdk==1.20.0
MarkupSafe==2.0.1
pymongo==3.12.0
pytz==2021.1
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
Werkzeug==2.0.1
zope.interface==5.4.0
